Added parameter_name and related fields
The parameter_name field is intended to be a human-readable,
representation of the "minimum" set of metadata fields that in
practice will often suffice for deciding what is measured by a
given time_series.

It will have the following form:

  <standard name>_<level>_<function>_<period>

Note that currently only standard_name is an official ACDD attribute.

Note also that two different time series may have the same
parameter_name (and differ e.g. in the instrument field).