Added partial support for reflectable int64s
This change adds partial support for reflectable metadata fields of type
int64, with the ts metadata field 'period' as the first example.
This is a basis for later supporting efficient range filtering.
The following two caveats apply for now: [1] the default int64 value of 0
is omitted from the output (since protoc puts 'omitempty' for all fields
in the generated file); [2] it is not possible to filter at all on
reflectable int64 fields. Both will be addressed in later commits.